requirements.txt
gunicorn
django
django-tinymce4-lite